// enums.ts
export enum MemberRole {
  TEACHER = "TEACHER",
  STUDENT = "STUDENT",
  LEADER = "LEADER",
  GUEST = "GUEST"
}

export enum ChannelType {
  TEXT = "TEXT",
  AUDIO = "AUDIO",
  VIDEO = "VIDEO"
}

export enum DeadlineStatus {
  UPCOMING = "UPCOMING",
  IN_PROGRESS = "IN_PROGRESS",
  SUBMITTED = "SUBMITTED"
}